Refact bricks.recurrent module
Split it into 3 submodules
