+++ blocks/bricks/recurrent/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseRecurrent, recurrent
+from .architectures import SimpleRecurrent, LSTM, GatedRecurrent
+from .misc import Bidirectional, RecurrentStack, RECURRENTSTACK_SEPARATOR
